move flavors to flavor
