+++ src/chains/ethereum/ethereum/package.json
@@ -89,7 +89,7 @@
     "ws": "8.2.3"
   },
   "devDependencies": {
-    "@ganache/flavors": "0.6.0",
+    "@ganache/flavor": "0.6.0",
     "@trufflesuite/uws-js-unofficial": "20.10.0-unofficial.1",
     "@types/abstract-leveldown": "7.2.0",
     "@types/encoding-down": "5.0.0",
