+++ src/chains/filecoin/filecoin/package.json
@@ -59,7 +59,7 @@
     "@filecoin-shipyard/lotus-client-rpc": "0.2.0",
     "@filecoin-shipyard/lotus-client-schema": "2.0.0",
     "@ganache/filecoin-options": "0.6.0",
-    "@ganache/flavors": "0.6.0",
+    "@ganache/flavor": "0.6.0",
     "@ganache/utils": "0.6.0",
     "@trufflesuite/uws-js-unofficial": "20.10.0-unofficial.1",
     "@types/abstract-leveldown": "7.2.0",
