+++ src/chains/filecoin/filecoin/tsconfig.json
@@ -7,14 +7,8 @@
     "strictNullChecks": true,
     "noImplicitAny": true
   },
-  "include": [
-    "index.ts",
-    "src/**/*"
-  ],
-  "typeRoots": [
-    "./node_modules/@types",
-    "./src/@types"
-  ],
+  "include": ["index.ts", "src/**/*"],
+  "typeRoots": ["./node_modules/@types", "./src/@types"],
   "references": [
     {
       "name": "@ganache/filecoin-options",
@@ -25,7 +19,7 @@
       "path": "../../../packages/utils"
     },
     {
-      "name": "@ganache/flavors",
+      "name": "@ganache/flavor",
       "path": "../../../packages/flavors"
     }
   ]
