+++ src/packages/cli/tsconfig.json
@@ -5,17 +5,14 @@
     "declarationDir": "typings",
     "composite": true
   },
-  "include": [
-    "index.ts",
-    "src/**/*.ts"
-  ],
+  "include": ["index.ts", "src/**/*.ts"],
   "references": [
     {
       "name": "@ganache/core",
       "path": "../core"
     },
     {
-      "name": "@ganache/flavors",
+      "name": "@ganache/flavor",
       "path": "../flavors"
     },
     {
