+++ src/packages/flavor/.npmignore
