+++ src/packages/flavor/LICENSE
