+++ src/packages/flavor/README.md
@@ -1,3 +1,3 @@
-# `@ganache/flavors`
+# `@ganache/flavor`
 
 Ganache's flavors enumeration and TypeScript types
