+++ src/packages/ganache/package.json
@@ -58,7 +58,7 @@
   "devDependencies": {
     "@ganache/cli": "0.6.0",
     "@ganache/core": "0.6.0",
-    "@ganache/flavors": "0.6.0",
+    "@ganache/flavor": "0.6.0",
     "@microsoft/api-extractor": "7.20.1",
     "@types/mocha": "9.0.0",
     "@types/node": "17.0.0",
